+++ web/android/app/src/main/java/io/github/daood40/quiz/MainActivity.java
@@ -1,0 +1,5 @@
+package io.github.daood40.quiz;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}